refactor(S-13): :truck: move type
